Part 3 Attempt
For the client I am having trouble trying to check if the packets received are in the correct order. I added a spot in the header (buffer) of each packet (in the server) to keep track of each packet (1-5). I am currently printing out which packets were received and what order they were received. For some reason it doesn't seem like we're getting it in the correct order. I commented my logic a bit.